fixed unsigned/signed comparison warning
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@3677 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/build.cpp
@@ -1319,7 +1319,7 @@ int CEXEBuild::resolve_jump_int(const char *fn, int *a, int offs, int start, int
       for (; jump != 0; jump -= direction)
       {
         offs += direction;
-        if (offs >= 0 && offs < cur_instruction_entry_map->getlen() * sizeof(int))
+        if (offs >= 0 && offs < cur_instruction_entry_map->getlen() * (int) sizeof(int))
         {
           while (skip_map[offs])
           {
